// fast-crm/src/triage-bot/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartMemory, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AGENT_MEMORY: SmartMemory;
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CRM_DATABASE: SmartSql;
  CRM_SERVICE: ServiceStub<import('../crm-service/index.js').default>;
  logger: Logger;
  RESPONSE_BOT: ServiceStub<import('../response-bot/index.js').default>;
  tracer: Tracer;
}